IOS-6207 Extract widget pre-warm into 3 prewarmers, spawn from VM init
