+++ Anytype/Sources/PresentationLayer/Modules/HomeWidgets/Container/HomeWidgetsViewModel.swift
@@ -22,20 +22,12 @@ final class HomeWidgetsViewModel {
     private var homeSectionsStorage: any HomeSectionsStorageProtocol
     @Injected(\.participantsStorage) @ObservationIgnored
     private var accountParticipantStorage: any ParticipantsStorageProtocol
-    @Injected(\.expandedService) @ObservationIgnored
-    private var expandedService: any ExpandedServiceProtocol
-    @Injected(\.setSubscriptionDataBuilder) @ObservationIgnored
-    private var setSubscriptionDataBuilder: any SetSubscriptionDataBuilderProtocol
-    @Injected(\.subscriptionStorageProvider) @ObservationIgnored
-    private var subscriptionStorageProvider: any SubscriptionStorageProviderProtocol
-    @Injected(\.spaceViewsStorage) @ObservationIgnored
-    private var spaceViewsStorage: any SpaceViewsStorageProtocol
-    @Injected(\.chatMessagesPreviewsStorage) @ObservationIgnored
-    private var chatMessagesPreviewsStorage: any ChatMessagesPreviewsStorageProtocol
-    @Injected(\.chatDetailsStorage) @ObservationIgnored
-    private var chatDetailsStorage: any ChatDetailsStorageProtocol
-    @Injected(\.objectsWithUnreadDiscussionsSubscription) @ObservationIgnored
-    private var unreadDiscussionsSubscription: any ObjectsWithUnreadDiscussionsSubscriptionProtocol
+    @Injected(\.setWidgetsPrewarmer) @ObservationIgnored
+    private var setWidgetsPrewarmer: any SetWidgetsPrewarmerProtocol
+    @Injected(\.treeWidgetsPrewarmer) @ObservationIgnored
+    private var treeWidgetsPrewarmer: any TreeWidgetsPrewarmerProtocol
+    @Injected(\.unreadSectionPrewarmer) @ObservationIgnored
+    private var unreadSectionPrewarmer: any UnreadSectionPrewarmerProtocol
 
     @ObservationIgnored
     weak var output: (any HomeWidgetsModuleOutput)?
@@ -55,6 +47,9 @@ final class HomeWidgetsViewModel {
     // matches pre-refactor `homeState = .readonly` default and `PinnedSectionViewModel`.
     private var canEdit: Bool = false
 
+    @ObservationIgnored
+    private var prewarmTask: Task<Void, Never>?
+
     var spaceId: String { info.accountSpaceId }
 
     var visibleSections: [HomeSection] {
@@ -73,31 +68,22 @@ final class HomeWidgetsViewModel {
     ) {
         self.info = info
         self.output = output
+        // Spawn pre-warm in init so it overlaps the space-transition animation,
+        // not from `.task` on view appear. Observation propagates the Prefetched*
+        // results to the view once `runPrewarm` populates state.
+        prewarmTask = Task(priority: .high) { [weak self] in
+            await self?.runPrewarm()
+        }
     }
 
-    func openWidgetObjects() async {
-        // Docs may still be opening (kicked off upstream in SpaceLoadingContainerViewModel);
-        // wait so prewarm sees a live channel doc.
-        await widgetsObjectsStorage.waitForReady(spaceId: info.accountSpaceId)
-        guard !Task.isCancelled,
-              let (channel, personal) = widgetsObjectsStorage.widgetsObjects(spaceId: info.accountSpaceId)
-        else { return }
+    deinit {
+        prewarmTask?.cancel()
+    }
 
-        // Pre-warm before flipping the section gate so Set/Type and expanded Tree
-        // widgets render rows on the first frame. We accept the loader extension in
-        // every context — a visible row-pop is worse than a small delay regardless
-        // of presentation.
-        async let prefetchedSet = prewarmSetWidgetSubscriptions(channelWidgetsObject: channel)
-        async let prefetchedTree = prewarmTreeWidgetChildren(channelWidgetsObject: channel)
-        async let prefetchedUnread = prewarmUnreadSection()
-        let (setMap, treeMap, unread) = await (prefetchedSet, prefetchedTree, prefetchedUnread)
-
-        guard !Task.isCancelled else { return }
-        channelWidgetsObject = channel
-        personalWidgetsObject = personal
-        prefetchedSetSubscriptions = setMap
-        prefetchedTreeChildren = treeMap
-        prefetchedUnreadSection = unread
+    /// Test synchronization hook — awaits the init-spawned pre-warm Task.
+    /// Not called from production code; the view reads state via Observation.
+    func awaitPrewarm() async {
+        await prewarmTask?.value
     }
 
     func startSubscriptions() async {
@@ -129,6 +115,30 @@ final class HomeWidgetsViewModel {
     }
 
     // MARK: - Private
+
+    private func runPrewarm() async {
+        // Docs may still be opening (kicked off upstream in ActiveSpaceManager);
+        // wait so prewarm sees a live channel doc.
+        await widgetsObjectsStorage.waitForReady(spaceId: info.accountSpaceId)
+        guard !Task.isCancelled,
+              let (channel, personal) = widgetsObjectsStorage.widgetsObjects(spaceId: info.accountSpaceId)
+        else { return }
+
+        // Pre-warm before flipping the section gate so Set/Type and expanded Tree
+        // widgets render rows on the first frame. The view body guards on
+        // `channelWidgetsObject != nil`; Observation re-renders once we assign below.
+        async let prefetchedSet = setWidgetsPrewarmer.prewarm(channelDoc: channel, spaceId: info.accountSpaceId)
+        async let prefetchedTree = treeWidgetsPrewarmer.prewarm(channelDoc: channel, spaceId: info.accountSpaceId)
+        async let prefetchedUnread = unreadSectionPrewarmer.prewarm(spaceId: info.accountSpaceId)
+        let (setMap, treeMap, unread) = await (prefetchedSet, prefetchedTree, prefetchedUnread)
+
+        guard !Task.isCancelled else { return }
+        channelWidgetsObject = channel
+        personalWidgetsObject = personal
+        prefetchedSetSubscriptions = setMap
+        prefetchedTreeChildren = treeMap
+        prefetchedUnreadSection = unread
+    }
 
     private func startSectionsConfigurationTask() async {
         for await configuration in homeSectionsStorage.configurationPublisher(spaceId: info.accountSpaceId).values {
@@ -188,231 +198,6 @@ final class HomeWidgetsViewModel {
             homepageTask = Task { [weak self] in
                 await self?.observeHomepageObject(objectId: next.objectId, canSetHomepage: next.canSetHomepage)
             }
-        }
-    }
-
-    // MARK: - Widget pre-warm
-
-    /// Per-widget budget. A slow widget falls back to its own mount-time open
-    /// (header first, rows later) instead of stalling the whole gate.
-    private static let prewarmTimeout: TimeInterval = 0.6
-
-    private func prewarmSetWidgetSubscriptions(
-        channelWidgetsObject: any BaseDocumentProtocol
-    ) async -> [String: PrefetchedSetSubscription] {
-        let setWidgets = channelWidgetsObject.children.compactMap { child -> BlockWidgetInfo? in
-            guard child.isWidget,
-                  let info = channelWidgetsObject.widgetInfo(block: child),
-                  info.isSetTypeWidget,
-                  expandedService.isExpanded(id: info.id, defaultValue: true)
-            else { return nil }
-            return info
-        }
-
-        guard setWidgets.isNotEmpty else { return [:] }
-
-        return await withTaskGroup(of: (String, PrefetchedSetSubscription)?.self) { group in
-            for widgetInfo in setWidgets {
-                group.addTask { [weak self] in
-                    guard let self else { return nil }
-                    guard let prefetched = await prewarmSingleSetWidget(widgetInfo: widgetInfo) else { return nil }
-                    return (widgetInfo.id, prefetched)
-                }
-            }
-
-            var result: [String: PrefetchedSetSubscription] = [:]
-            for await pair in group {
-                if let (id, prefetched) = pair {
-                    result[id] = prefetched
-                }
-            }
-            return result
-        }
-    }
-
-    /// In-memory snapshots — no timeout; cold-start aggregator returns `[:]` and the live subscription fills gaps.
-    private func prewarmUnreadSection() async -> PrefetchedUnreadSection? {
-        let spaceView = spaceViewsStorage.spaceView(spaceId: spaceId)
-        let supportsMultiChats = !(spaceView?.isOneToOne ?? false)
-        guard supportsMultiChats else { return nil }
-
-        async let previews = chatMessagesPreviewsStorage.previews()
-        async let chats = chatDetailsStorage.allChats()
-        async let unreadBySpace = unreadDiscussionsSubscription.snapshot
-
-        let (p, c, u) = await (previews, chats, unreadBySpace)
-        guard let spaceView else { return PrefetchedUnreadSection(rows: [], supportsMultiChats: true) }
-
-        let rows = UnreadSectionViewModel.mergeUnreadRows(
-            previews: p,
-            chatDetails: c,
-            spaceView: spaceView,
-            unreadBySpace: u,
-            spaceId: spaceId
-        )
-        return PrefetchedUnreadSection(rows: rows, supportsMultiChats: true)
-    }
-
-    private func prewarmTreeWidgetChildren(
-        channelWidgetsObject: any BaseDocumentProtocol
-    ) async -> [String: PrefetchedTreeChildren] {
-        let treeWidgets = channelWidgetsObject.children.compactMap { child -> BlockWidgetInfo? in
-            guard child.isWidget,
-                  let info = channelWidgetsObject.widgetInfo(block: child),
-                  info.isTreeObjectWidget,
-                  expandedService.isExpanded(id: info.id, defaultValue: true)
-            else { return nil }
-            return info
-        }
-
-        guard treeWidgets.isNotEmpty else { return [:] }
-
-        return await withTaskGroup(of: (String, PrefetchedTreeChildren)?.self) { group in
-            for widgetInfo in treeWidgets {
-                group.addTask { [weak self] in
-                    guard let self else { return nil }
-                    guard let prefetched = await prewarmSingleTreeWidget(widgetInfo: widgetInfo) else { return nil }
-                    return (widgetInfo.id, prefetched)
-                }
-            }
-
-            var result: [String: PrefetchedTreeChildren] = [:]
-            for await pair in group {
-                if let (id, prefetched) = pair {
-                    result[id] = prefetched
-                }
-            }
-            return result
-        }
-    }
-
-    private func prewarmSingleTreeWidget(widgetInfo: BlockWidgetInfo) async -> PrefetchedTreeChildren? {
-        guard case let .object(linkedObjectDetails) = widgetInfo.source else { return nil }
-
-        // Target has no children — seed an empty list so the widget renders `.empty`
-        // synchronously on first frame instead of flashing `.loading` → `.empty`
-        // when the live subscription emits its first (empty) result.
-        guard linkedObjectDetails.links.isNotEmpty else {
-            return PrefetchedTreeChildren(childDetails: [])
-        }
-
-        return await withTimeout(seconds: Self.prewarmTimeout) { [self] in
-            // Fresh builder per widget — its `subscriptionId` is the storage's `subId`,
-            // and we need a unique pair so disposable storages don't collide.
-            let builder = TreeSubscriptionDataBuilder()
-            let storage = subscriptionStorageProvider.createSubscriptionStorage(
-                subId: builder.subscriptionId
-            )
-            let subscriptionData = builder.build(
-                spaceId: info.accountSpaceId,
-                objectIds: linkedObjectDetails.links
-            )
-
-            do {
-                try await storage.startOrUpdateSubscription(data: subscriptionData)
-            } catch {
-                return nil
-            }
-
-            // `statePublisher` is backed by CurrentValueSubject, so subscribing after
-            // `startOrUpdateSubscription` returns replays the current state immediately.
-            var snapshot: [ObjectDetails]?
-            for await state in storage.statePublisher.values {
-                snapshot = state.items
-                break
-            }
-            try? await storage.stopSubscription()
-
-            guard let items = snapshot else { return nil }
-
-            // Mirror `TreeSubscriptionManager`'s publisher composition: filter to
-            // openable objects, sort by parent `.links` order, then apply the limit.
-            let sorted = items
-                .filter(\.isNotDeletedAndSupportedForOpening)
-                .reordered(by: linkedObjectDetails.links, transform: { $0.id })
-            return PrefetchedTreeChildren(childDetails: Array(sorted.prefix(widgetInfo.fixedLimit)))
-        }
-    }
-
-    private func prewarmSingleSetWidget(widgetInfo: BlockWidgetInfo) async -> PrefetchedSetSubscription? {
-        guard case let .object(setDetails) = widgetInfo.source else { return nil }
-
-        return await withTimeout(seconds: Self.prewarmTimeout) { [self] in
-            let setDocument = documentsProvider.setDocument(
-                objectId: setDetails.id,
-                spaceId: info.accountSpaceId,
-                mode: .preview
-            )
-
-            do { try await setDocument.open() } catch { return nil }
-
-            // `setDocument.updateData()` runs on the next main-queue tick after open()
-            // returns (syncPublisher replays via `receiveOnMain`). Check the actual
-            // state on every emission — `setUpdatePublisher` can emit `.syncStatus`
-            // before `.dataviewUpdated`, and we'd loop without progress if we only
-            // matched on the event variant.
-            if !setDocument.dataView.views.isNotEmpty {
-                for await _ in setDocument.setUpdatePublisher.values {
-                    if setDocument.dataView.views.isNotEmpty { break }
-                }
-            }
-
-            guard setDocument.canStartSubscription(), setDocument.dataView.views.isNotEmpty else {
-                return nil
-            }
-
-            let subscriptionData = buildSubscriptionData(
-                widgetInfo: widgetInfo,
-                setDocument: setDocument
-            )
-
-            let storage = subscriptionStorageProvider.createSubscriptionStorage(
-                subId: subscriptionData.identifier
-            )
-            do {
-                try await storage.startOrUpdateSubscription(data: subscriptionData)
-            } catch {
-                return nil
-            }
-
-            // `statePublisher` is backed by CurrentValueSubject, so subscribing after
-            // `startOrUpdateSubscription` returns replays the current state immediately.
-            for await state in storage.statePublisher.values {
-                return PrefetchedSetSubscription(
-                    setDocument: setDocument,
-                    subscriptionStorage: storage,
-                    state: state
-                )
-            }
-            return nil
-        }
-    }
-
-    private func buildSubscriptionData(
-        widgetInfo: BlockWidgetInfo,
-        setDocument: any SetDocumentProtocol
-    ) -> SubscriptionData {
-        setSubscriptionDataBuilder.widgetSubscriptionData(
-            widgetInfo: widgetInfo,
-            setDocument: setDocument,
-            identifier: "SetWidget-\(UUID().uuidString)",
-            spaceType: spaceViewsStorage.spaceView(spaceId: setDocument.spaceId)?.spaceType
-        )
-    }
-
-    private func withTimeout<T: Sendable>(
-        seconds: TimeInterval,
-        operation: @escaping @MainActor () async -> T?
-    ) async -> T? {
-        await withTaskGroup(of: T?.self) { group in
-            group.addTask { @MainActor in await operation() }
-            group.addTask {
-                try? await Task.sleep(for: .seconds(seconds))
-                return nil
-            }
-            defer { group.cancelAll() }
-            // `next()` returns Element? = T??; the `?? nil` flattens to T?.
-            return await group.next() ?? nil
         }
     }
